+++ trane/__init__.py
@@ -7,6 +7,8 @@ from trane.datasets import (
     load_youtube_metadata,
     load_yelp,
 )
+from trane.column_schema import ColumnSchema
+from trane.logical_types import *
 from trane.utils import *  # noqa
 from trane.version import __version__